some renaming + exception not baseexception
