+++ src/inspect_ai/model/_providers/anthropic.py
@@ -319,7 +319,7 @@ class AnthropicAPI(ModelAPI):
         return str(self.api_key)
 
     @override
-    def should_retry(self, ex: BaseException) -> bool:
+    def should_retry(self, ex: Exception) -> bool:
         if isinstance(ex, APIStatusError):
             return is_retryable_http_status(ex.status_code)
         elif isinstance(ex, APITimeoutError):
